refactor: styles folder
